Initial commit:
Include function that will parse the file passed as arguments for <parm> tags and populate the variables of the mainscript with matching variable names.

parameters should be contained within <head><parm></parm></head> of the uhp template.

parameters should follow the following format:

<_varname_-_x idx_-_y idx_>value</_varname_>
if the value is to be mapped to a simple variable the x and y idx can be omitted.

git-svn-id: https://usc-uhp.googlecode.com/svn/trunk@17 f1e605ba-d566-c5ae-5db7-88967aac56d5